Conflicts:
	TP3/src/ca/csf/RTS/game/entity/controllableEntity/building/WatchTower.java

+++ TP3/src/ca/csf/RTS/game/entity/state/Training.java
@@ -25,8 +25,8 @@ public class Training implements State {
 		}
 	}
 	
-	public float getPourcentageDone(){
-		return timeElapsed/ trainer.getNextInQueue().time() * 100;
+	public int getPourcentageDone(){
+		return (int) (timeElapsed/ trainer.getNextInQueue().time() * 100);
 	}
 	
 	public String getStateName() {
